Stub for generating the ics
